refactor: move alert icon state to common package to share betwen live and analysis UI
