+++ src/ui/live/telemetry_view.rs
@@ -2,7 +2,8 @@ use std::sync::Arc;
 
 use egui::{Color32, CornerRadius, Frame, Id, ImageButton, Layout, Sense, Vec2b, ViewportCommand};
 use egui_plot::{Line, PlotPoints};
-use log::debug;
+
+use crate::ui::stroke_shade;
 
 use super::{
     LiveTelemetryApp, DEFAULT_BUTTON_CORNER_RADIUS, DEFAULT_WINDOW_CORNER_RADIUS, PALETTE_ORANGE,
@@ -31,13 +32,15 @@ impl LiveTelemetryApp {
                     ui.add_space(10.);
                     // icons from https://remixicon.com/
                     ui.add(
-                        ImageButton::new(egui::include_image!("../../assets/tools-fill.png"))
+                        ImageButton::new(egui::include_image!("../../../assets/tools-fill.png"))
                             .corner_radius(DEFAULT_BUTTON_CORNER_RADIUS),
                     );
                     if ui
                         .add(
-                            ImageButton::new(egui::include_image!("../../assets/alert-fill.png"))
-                                .corner_radius(DEFAULT_BUTTON_CORNER_RADIUS),
+                            ImageButton::new(egui::include_image!(
+                                "../../../assets/alert-fill.png"
+                            ))
+                            .corner_radius(DEFAULT_BUTTON_CORNER_RADIUS),
                         )
                         .clicked()
                     {
@@ -49,7 +52,7 @@ impl LiveTelemetryApp {
                         if ui
                             .add(
                                 ImageButton::new(egui::include_image!(
-                                    "../../assets/close-circle-fill.png"
+                                    "../../../assets/close-circle-fill.png"
                                 ))
                                 .corner_radius(DEFAULT_BUTTON_CORNER_RADIUS),
                             )
@@ -126,24 +129,4 @@ impl LiveTelemetryApp {
         // make it always repaint. TODO: can we slow down here?
         ctx.request_repaint();
     }
-}
-
-pub(crate) fn stroke_shade(start: Color32, end: Color32, y: f32) -> Color32 {
-    Color32::from_rgb(
-        u8::try_from(
-            (start.r() as f32 + y * (end.r() as f32 - start.r() as f32)).clamp(0., 255.) as u32,
-        )
-        .map_err(|e| debug!("Error interpolating colors: {}", e))
-        .unwrap(),
-        u8::try_from(
-            (start.g() as f32 + y * (end.g() as f32 - start.g() as f32)).clamp(0., 255.) as u32,
-        )
-        .map_err(|e| debug!("Error interpolating colors: {}", e))
-        .unwrap(),
-        u8::try_from(
-            (start.b() as f32 + y * (end.b() as f32 - start.b() as f32)).clamp(0., 255.) as u32,
-        )
-        .map_err(|e| debug!("Error interpolating colors: {}", e))
-        .unwrap(),
-    )
 }
